Comments for XML-Files, security check on android and minor bugfixes
Added a comment to the generated XML-files that reads that neither file-content nor file-name should be modified. Also the security check (for which the filename is compared to the name attribute of the root tag) has been added again in android, since it was not working.
Furthermore there were minor bugs that were fixed.